refactor assignments page
